Checkpoint new structure
