+++ backend/src/auth/dto/index.ts
@@ -4,5 +4,3 @@ export * from './password-reset.dto';
 export * from './register-brand.dto';
 export * from './brand-registration-response.dto';
 export * from './create-brand.dto';
-export * from './validation-response.dto';
-export * from './validation-request.dto';
