+++ backend/src/validate/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './validate-email.dto';
+export * from './validate-username.dto';
