+++ backend/src/validate/validate.controller.ts
@@ -1,0 +1,46 @@
+import { Controller, Post, Body, ValidationPipe, HttpCode, HttpStatus } from '@nestjs/common';
+import { ApiTags, ApiOperation, ApiResponse, ApiBody } from '@nestjs/swagger';
+import { ValidateService } from './validate.service';
+import { 
+  ValidateEmailDto, 
+  ValidateUsernameDto, 
+  EmailValidationResponseDto, 
+  UsernameValidationResponseDto 
+} from './dto';
+import { BaseResponseDto } from '../common/dto';
+
+@ApiTags('Validación')
+@Controller('validate')
+export class ValidateController {
+  constructor(private readonly validateService: ValidateService) {}
+
+  @Post('email')
+  @HttpCode(HttpStatus.OK)
+  @ApiOperation({ summary: 'Validar disponibilidad de email' })
+  @ApiBody({ type: ValidateEmailDto })
+  @ApiResponse({ 
+    status: 200, 
+    description: 'Email validado exitosamente',
+    type: BaseResponseDto<EmailValidationResponseDto>
+  })
+  async validateEmail(
+    @Body(ValidationPipe) validateEmailDto: ValidateEmailDto
+  ): Promise<BaseResponseDto<EmailValidationResponseDto>> {
+    return this.validateService.validateEmail(validateEmailDto.email, validateEmailDto.brandId);
+  }
+
+  @Post('username')
+  @HttpCode(HttpStatus.OK)
+  @ApiOperation({ summary: 'Validar disponibilidad de username' })
+  @ApiBody({ type: ValidateUsernameDto })
+  @ApiResponse({ 
+    status: 200, 
+    description: 'Username validado exitosamente',
+    type: BaseResponseDto<UsernameValidationResponseDto>
+  })
+  async validateUsername(
+    @Body(ValidationPipe) validateUsernameDto: ValidateUsernameDto
+  ): Promise<BaseResponseDto<UsernameValidationResponseDto>> {
+    return this.validateService.validateUsername(validateUsernameDto.username);
+  }
+}
